requirements.txt
setuptools==58.2.0
aiobotocore==1.0.4
aiohttp==3.7.4.post0
aioitertools==0.7.0
alabaster==0.7.12
appnope==0.1.0
argcomplete==1.9.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
bleach==3.1.4
boto3==1.12.22
botocore==1.15.32
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
durationpy==0.5
entrypoints==0.3
firebase-admin==3.0.0
firestore==0.0.8
flake8==3.7.8
Flask==1.1.1
Flask-Caching==1.7.2
Flask-Cors==3.0.8
future==0.16.0
gitdb==4.0.4
GitPython==3.1.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
gunicorn==20.0.4
hjson==3.0.1
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
imagesize==1.1.0
importlib-metadata==0.22
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kappa==0.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.2
multidict==4.7.6
mypy==0.790
mypy-extensions==0.4.3
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.6
nltk==3.4.5
notebook==6.0.3
numpy==1.17.2
packaging==19.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
placebo==0.9.0
pluggy==0.13.0
portalocker==1.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.15.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.16.0
pytest==5.1.2
pytest-dotenv==0.5.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.1.2
pyzmq==19.0.0
redis==3.4.1
requests==2.22.0
rsa==4.0
s3transfer==0.3.3
sacrebleu==1.4.10
Send2Trash==1.5.0
six==1.12.0
smmap==3.0.2
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.19.1
traitlets==4.3.3
troposphere==3.0.3
typed-ast==1.4.0
typing-extensions==3.10.0.2
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
wrapt==1.12.1
wsgi-request-logger==0.4.6
yarl==1.4.2
zipp==0.6.0
scikit-learn==0.24.2
fastapi==0.63.0
uvicorn==0.13.4
starlette==0.13.6
pydantic==1.8.1
fastapi-cache2==0.1.8
types-cachetools==5.2.1
sentry-sdk==1.9.0